# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "description": "Cria um novo evento para a comunidade do usuário autenticado",
        "operationId": "EventController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "403": {
            "description": "Usuário não tem permissão de comunidade"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar novo evento",
        "tags": [
          "events"
        ]
      },
      "get": {
        "description": "Lista eventos com paginação HATEOAS. Pode filtrar por comunidade, modalidade e status",
        "operationId": "EventController_getAll",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros",
            "schema": {
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros para pular",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "communityId",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID da comunidade",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "modality",
            "required": false,
            "in": "query",
            "description": "Filtrar por modalidade",
            "schema": {
              "enum": [
                "ONLINE",
                "PRESENCIAL",
                "HIBRIDO"
              ],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos retornada com sucesso com links de navegação"
          }
        },
        "summary": "Listar todos os eventos",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "operationId": "EventController_getByID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do evento"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "summary": "Buscar evento por ID",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "description": "Atualiza um evento existente. Apenas o dono da comunidade pode atualizar seus eventos",
        "operationId": "EventController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idAddress",
            "required": false,
            "in": "query",
            "description": "ID do endereço (se estiver atualizando)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso"
          },
          "403": {
            "description": "Sem permissão para atualizar este evento"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar evento",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "description": "Remove um evento. Apenas o dono da comunidade pode deletar seus eventos",
        "operationId": "EventController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento deletado com sucesso"
          },
          "403": {
            "description": "Sem permissão para deletar este evento"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar evento",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/tickets": {
      "post": {
        "description": "Cria um novo tipo de ticket para eventos da comunidade",
        "operationId": "TicketController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket criado com sucesso"
          },
          "403": {
            "description": "Usuário não tem permissão de comunidade"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar novo ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "description": "Lista tipos de tickets disponíveis com paginação HATEOAS e filtros",
        "operationId": "TicketController_getAll",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros",
            "schema": {
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros para pular",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID do evento",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tickets retornada com sucesso com links de navegação"
          }
        },
        "summary": "Listar todos os tickets",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}": {
      "get": {
        "operationId": "TicketController_getByID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do ticket"
          },
          "404": {
            "description": "Ticket não encontrado"
          }
        },
        "summary": "Buscar ticket por ID",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "description": "Atualiza um tipo de ticket. Apenas o dono da comunidade pode atualizar",
        "operationId": "TicketController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idAddress",
            "required": false,
            "in": "query",
            "description": "ID do endereço (se aplicável)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket atualizado com sucesso"
          },
          "403": {
            "description": "Sem permissão para atualizar este ticket"
          },
          "404": {
            "description": "Ticket não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar ticket",
        "tags": [
          "tickets"
        ]
      },
      "delete": {
        "description": "Remove um tipo de ticket. Apenas o dono da comunidade pode deletar",
        "operationId": "TicketController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket deletado com sucesso"
          },
          "403": {
            "description": "Sem permissão para deletar este ticket"
          },
          "404": {
            "description": "Ticket não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar ticket",
        "tags": [
          "tickets"
        ]
      }
    }
  },
  "info": {
    "title": "EventDev API",
    "description": "API RESTful para gerenciamento de eventos e comunidades com autorização baseada em papéis e permissões",
    "version": "1.0",
    "contact": {
      "name": "EventDev Team",
      "url": "https://eventdev.org",
      "email": "contact@eventdev.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5122",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Token de autenticação JWT do SuperTokens",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}